+++ include/model.h
@@ -6,10 +6,10 @@
 typedef struct model_t{
     layer_t** layers; //array of pointers to layers
     int num_layers;
-    float learning_rate;
+    double learning_rate;
 }model_t;
 
-model_t* model_init(int input_size, int* layer_sizes, enum activation_type* act, int num_layers, float learning_rate);
+model_t* model_init(int input_size, int* layer_sizes, enum activation_type* act, int num_layers, double learning_rate);
 matrix_t* model_predict(model_t* model, matrix_t* input);
 void model_train(model_t* model, matrix_t* x, matrix_t* y, int epochs);
 void model_free(model_t* model);
