+++ src/debug.c
@@ -1,6 +1,6 @@
 #include "../include/debug.h"
 
-void matrix_print_rowcol(float* arr, int size,int flag){
+void matrix_print_rowcol(double* arr, int size,int flag){
     int i = 0;
     for(;i<size;++i){
         printf("[%.2f]",arr[i]);
@@ -17,7 +17,7 @@ void matrix_print(const matrix_t *m){
     const int cols = m->cols;
     const int mat_size = m->rows * cols;
     int i = 0;
-    float* d = m->data;
+    double* d = m->data;
     for(;i<mat_size;++i){
         if(i%cols == 0){
             printf("\n");
@@ -31,8 +31,8 @@ void matrix_print(const matrix_t *m){
 void matrix_set_linear_range(matrix_t* m){
     int i = 0;
     const int num_elem = m->cols * m->rows;
-    float* arr = m->data;
+    double* arr = m->data;
     for(; i<num_elem; ++i){
-        arr[i] = (float)i;
+        arr[i] = (double)i;
     }
 }
